refactor(esbuild-plugin-gjsify): extract node_modules path rewrite into shared util
The onLoad hooks in app/gjs.ts and app/node.ts had ~40 lines of
near-identical logic for rewriting `import.meta.url`,
`__dirname`, and `__filename` in node_modules files. A bug fix in
one branch (e.g. the recent loader detection extension) had to be
replicated by hand in the other; that is exactly the maintenance
trap this commit removes.

Both targets now import `registerNodeModulesPathRewrite` from
`utils/rewrite-node-modules-paths.ts`, which:
- detects ESM (has `import.meta.url`) vs CJS files
- ESM: replaces `import.meta.url` with `new URL(relPath, import.meta.url).href`
  and synthesises an ESM-safe `__dirname`/`__filename` preamble
- CJS: injects absolute `__dirname`/`__filename` string literals
  (introducing `import.meta.url` would make esbuild treat the
  module as ESM, breaking module.exports/require — typescript.js)
- preserves the loader:'ts' detection for .mts/.cts/.tsx sources

Behavior is identical to the previous duplicated implementations.
Verified Node 276/276 + GJS 212/212 (3 ignored) unchanged.
